Remove nokogiri dependency; process JSON Solr schemas
The nokogiri dependency is somewhat difficult to install on aarch64
(Arm) platforms, limiting the utility of this library. Solr now
supports JSON-based schemas, which (a) can be handled via the stdlib
and (b) are a little easier to parse.

Additionally, this makes some improvements to the Solr validation
process, such as checking for required fields and slightly
clearer semantics.